test: Move main test-runner into its own file
Create a new test.cpp file to hold the main function.
